Remove RuboCop
Github warns about a low security vulernability in the rubocop gem
(which is only used during development). I tried briefly to upgrade the
gem, but I was greeted with hundreds of new style checks. The vast
majority for whom RuboCop demanded the opposite in previous versions.

I'm just tired of spending my time to follow the guidelines of whoever
is in charge of RuboCop this year. As much as I'd like to enforce a
consistent code style, the rubocop project doesn't seem to have a clear
idea on such style either.