Require declaration of labels when creating metrics
Instead of allowing the first observation / setting of a metric to have
any labels, and then validating that any further observations match those,
we now require declaring what labels will be specified at the time of
instantiating the metric.

The Validator now only needs to verify that the provided labels match the
expectation, without needing to keep track of the first observation. For
now, it's still caching the validations, for performance, but that should
disappear in a soon-to-come commit.

`base_labels` now works slightly different. Instead of being a "base"
value that is set when creating the metric, and then combined with the
"other" labels at the time of exporting, it's now called `preset_values`,
and it that get merged as part of the labels stored with every observation.

This means the storage that holds the values will always have all the labels,
including the common ones. This allows having a method like the `labels()`
one recommended by the best practices, to have a "view" of a metric with
some labels pre-applied. This will be added on a future commit, once we
have a central data store.

Finally, this needed a little adaptation of the Collector middleware
that comes bundled in. This Collector automatically tracks metrics with
some pre-defined labels, but allows the user to define which labels
they want, on a per-request basis, by specifying a `proc` that will generate
the labels given a Rack `env`.

Given this design, we can't know what labels the custom `proc` will return
at the time of initializing the metric, and we don't have an `env` to
pass to them, so the interface of these procs has changed such that they
now need to be able to handle an empty `env`, and still return a `hash`
with all the right keys.

Signed-off-by: Daniel Magliola <danielmagliola@gocardless.com>