Abstract away storage of Metric data into DataStore objects
Currently, each Metric object stores its data (labels and values) in a Hash
object internal to itself. This change introduces the concept of a Data Store
that is external to the metrics themselves, which holds all the data.

The main reason to do this is that by having a standardized and simple
interface that metrics use to access the store, we abstract away the details
of storing the data from the specific needs of each metric.

This allows us to then simply swap around the stores based on the needs
of different circumstances, with no changes to the rest of the client. The
main use case for this is solving the "pre-fork" server scenario, where
multiple processes need to share the metric storage, to be able to report
coherent numbers to Prometheus.

This change provides one store, a Synchronized hash, that pretty much has
the same characteristics that the existing Storage had, plus some Mutex overhead.
Within a single process scenario, this should be the fastest way to operate in a
multi-threaded environment, and the safest. As such, this is the default store.

Future commits will introduce new ones, for specific scenarios, and also each
consumer of the Prometheus Client can make their own and simply swap them for
the built-in ones, if they have specific needs.

The interface and requirements of Stores are specified in detail in a README.md
file in the `client/data_stores` directory. This is the documentation that must
be used by anyone wishing to create their own store.

Signed-off-by: Daniel Magliola <danielmagliola@gocardless.com>