Short-circuit label validation on Metric observations
Every time a metric gets observed / incremented, we are composing the
labels hash, merging the labels that were just passed in as part of the
observation with the metric's `preset_labels`, and validating that the
labelset is valid.

If all labels are pre-set already, however (either as `preset_labels` when
declaring the metric, or by use of `with_labels`), we can validate when
the labels are pre-set, and then skip the validation on every observation,
which can lead to some decent time savings if a metric is observed often.

Signed-off-by: Daniel Magliola <danielmagliola@gocardless.com>