Accumulate Histogram Buckets on Export, not Observe
In a Histogram, each bucket's count includes also the counts of all
buckets smaller than it.

The original code for Histograms would handle this at `observe` time,
by incrementing all buckets bigger than the observed sample.

Since calls to the Data Store may be expensive, and also all these
increments need to be done atomically to guarantee the export doesn't
have inconsistent values, this can have a big effect on performance.

With this change, at `observe` time, we only increase the bucket the
sample falls in, and the `sum`, and do the accumulation when reading
the Histogram, which makes observations about 30% to 40% faster with a
"uniform" distribution of sample. A sample skewed towards smaller buckets
should show a much larger performance gain.

With this change, we also remove the `count` key from the result of
calling `get`. This key is redundant (it's the same value as `+Inf`,
and rendering this as `count` is the job of the Text Exporter already,
`get` shouldn't be returning it.