Remove support for flexible labels in collector middleware
We decided in #111 that the current interface for this is confusing to
the point where it's more trouble than it's worth.

The alternatives we came up with result in the middleware doing almost
nothing and delegating to user-provided lambdas.

Since we're pre-1.0, we're removing this in the belief that if it turns
out to be a widely-requested feature, we can come up with something
better. Leaving this implementation in would commit us to an interface
we don't like.

Fixes #111

Signed-off-by: Chris Sinjakli <chris@gocardless.com>