Make it clear when error comes from mocking in tests
This is a small quality of life enhancement when debugging test
failures. When you comment out the catch-all `rescue` in the production
code of Prometheus::Middleware::Collector, you see a `NoMethodError`
with no stack or message attached.

Initially, this is pretty confusing, and you may think it was the reason
for your tests failing! In fact, it's to test the catch-all rescue.

This commit makes it really explicit that the error is deliberately
coming from the tests.

Signed-off-by: Chris Sinjakli <chris@gocardless.com>