Update example README for #111
I was trying to modify the path that's exported and didn't realize it was _just_ changed. The documentation in the example project still reflects the old way.

Gave a short example of how I did it -- I realize subclassing a protected method is dangerous but there are no other hooks.

FWIW my use case was that the default regular expressions did not cover my url structure and I needed to add another one. I did not need to add more labels.

Signed-off-by: Sean Walberg <sean@ertw.com>