Remove `concurrent-ruby` dependency
We added this dependency to be able to use `ReadWriteLock`, which is
a reentract lock. We did not need the Read/Write locking distinction,
and at the time, we didn't know that `Monitor` was reentrant, so this
seemed to be the best solution.

Knowing that `Monitor` is reentrant, we can get rid of the dependency,
and simply use `Monitor` instead.

Note that we keep `concurrent-ruby` as a *dev* dependency, because the
performance benchmarks use other primitives provided by it to synvhronize
their threads, but it's not needed for specs to pass, or for production use.

Signed-off-by: Daniel Magliola <danielmagliola@gocardless.com>