Remove Labelset Validator Cache
Currently, when we validate a labelset on increment, we cache
that we've seen this particular labelset, so we can avoid
revalidating it next time we see it.

This is currently not thread-safe, since we're not synchronizing
access to the has we use for that cache. This is only a problem
in non-MRI, but we are supporting those.

Removing this cache has a slight performance hit for metrics
with labels, but removing the cache, in conjunction with the next
commit in this PR, makes the code *way* faster than the current
version, while simplifying the code, so we're opting for that.

The other alternative would have veen to add a Mutex around that
Hash. However, that makes it slower than the original, even if we
do the change from the next commit, so we're going for the simpler,
much faster, much more obvious path of not caching this validation.

This is probably also going to have a positive effect on memory usage.
That cache is definitely not free.

Signed-off-by: Daniel Magliola <danielmagliola@gocardless.com>