Avoid revalidating labelset keys on each metric observation
When declaring a metric, we declare what label keys will be used for it.
At that point, we validate that those are all valid keys (symbols,
match a given regex, etc).

Then, on each observation of the metric, we validate that the keys passed
in for the labels match the ones we were originally expecting, to make
sure all of those labels were set, but no others.

We were also, at that point, validating that the keys passed in are valid.
This validation is pretty slow, and it's redundant, since keys that aren't
valid won't match the expected ones anyway, so we can just compare just
that those match. This has quite a pronounced effect on performance.

Signed-off-by: Daniel Magliola <danielmagliola@gocardless.com>