Make DataStores shared examples enforce floats for metric values
The interface of a store is a labelset (hash of hashes) to a double.
It's important that we check the values are doubles rather than
integers. `==`, which is what `eq` calls allows conversion between
floats and integers (i.e. `5 == 5.0`). `eql` enforces that the two
numbers are of the same type.

Signed-off-by: Chris Sinjakli <chris@gocardless.com>