Add :most_recent aggregation to DirectFileStore
This reports the value that was set by a process most recently.

The way this works is by tagging each value in the files with the
timestamp of when they were set.

For all existing aggregations, we ignore that timestamp and do
what we've been doing so far.
For `:most_recent`, we take the "maximum" entry according to
its timestamp (i.e. the latest) and then return its value

Signed-off-by: Stefan Sundin <stefan@stefansundin.com>
Signed-off-by: Daniel Magliola <danielmagliola@gocardless.com>