Do now allow `DirectFileStore#increment` when using :most_recent aggregation
If we do this, we'd be incrementing the value for *this* process, not the
global one, which is almost certainly not what the user wants to do.

This is not very pretty because we may end up raising an exception in
production (as test/dev tend to not use DirectFileStore), but we consider
it better than letting the user mangle their numbers and end up with
incorrect metrics.

Signed-off-by: Daniel Magliola <danielmagliola@gocardless.com>