Add a method to clear the registry
Clearing the registry is something we would typically do in development
and test environments.

For instance, in Rails apps, this method can be used when reloading
classes so that auto-loaded files that register Prometheus metrics can
be reloaded:

```rb
Rails.application.reloader.after_class_unload do
  Prometheus::Client.registry.clear
end
```

Not clearing the registry before classes are reloaded would result in
the registry raising an `AlreadyRegisteredError` when a file
registering a metric is reloaded.

Signed-off-by: Matthieu Prat <matthieuprat@gocardless.com>