Include SCRIPT_NAME when determining path
When determining the path for a request, `Rack::Request` prefixes the
`SCRIPT_NAME`, [as seen here][1]. This makes sure the relative path is
included. At work, we use [Iodine](https://github.com/boazsegev/iodine) and
Turbolinks, and without this patch most of our requests paths in Promethus are
`""`.  This patch fixes that.

[1]: https://github.com/rack/rack/blob/294fd239a71aab805877790f0a92ee3c72e67d79/lib/rack/request.rb#L512

Signed-off-by: Ian Ker-Seymer <i.kerseymer@gmail.com>