Include SCRIPT_NAME when determining path in Collector
When determining the path for a request, `Rack::Request` prefixes the
`SCRIPT_NAME`, [as seen here][1].

This is a problem with our current code when using mountable engines,
where the engine part of the path gets lost.

This patch fixes that to include `SCRIPT_NAME` as part of the path.

NOTE: This is not backwards compatible. Labels will change in existing
metrics. We will cut a new major version once we ship this.

[1]: https://github.com/rack/rack/blob/294fd239a71aab805877790f0a92ee3c72e67d79/lib/rack/request.rb#L512

Co-authored-by: Ian Ker-Seymer <i.kerseymer@gmail.com>
Co-authored-by: Ruslan Kornev <oganer@gmail.com>
Signed-off-by: Daniel Magliola <dmagliola@crystalgears.com>