Automatically initialize time series without labels
According to [Prometheus Best Practices](https://prometheus.io/docs/practices/instrumentation/#avoid-missing-metrics),
client libraries are expected to automatically export a 0 value when declaring
a metric that has no labels.

We missed this so far in the Ruby client, this PR rectifies that.

NOTE: This can be considered a breaking change. On the one hand, it's a bug fix,
but on the other, it will make many time series materialize when scraping apps
that use the client, that previously wouldn't have.

Depending on how many label-less metrics the app is declaring, this may have a
significant impact, so we should probably cut a new major version and warn
about this in the Release Notes.

Signed-off-by: Daniel Magliola <dmagliola@crystalgears.com>