Add tests to `with_labels`
This reproduces the error reported in issue #225, where `with_labels`
doesn't work at all the way it is intended, and in addition to that,
if you're using DirectFileStore, it can very easily corrupt your files.

The problem is the new metric received after calling `with_labels` is a
copy of the original, with the labels pre-set, but it has
**its own internal store**, so it doesn't actually impact the values of
the original metric, which is what gets exported by the registry.

Signed-off-by: Daniel Magliola <dmagliola@crystalgears.com>