Explain the test that checks for file corruption
The test that reproduces our file corruption issue follows a very specific
set of steps to actually corrupt the files, and it's probably impossible
to figure out why.

Explaining those specific steps, however, requires explaining a lot about
the internals of `FileMappedDict`, of the files we store, and how corruption
happens in the first place, before we can understand why we need those
specific steps to reproduce it.

I hope this explanation makes sense.

Signed-off-by: Daniel Magliola <dmagliola@crystalgears.com>