Use base64 encoding when necessary in grouping key in push client
Certain label values (empty string, and anything containing a `/`)
cannot be used in the grouping key without extra work to encode them
properly.

Labels are represented as pairs of path segments in the URL, so an
unencoded `/` would be treated as a path separator. The empty string
would result in two consecturive path separators (`//`), which HTTP
libraries, proxies, and web servers are liable to normalise away.

This commit uses the base64 encoding method supported by the pushgateway
server to encode such values. See:

https://github.com/prometheus/pushgateway/blob/6393a901f56d4dda62cd0f6ab1f1f07c495b6354/README.md#url
Signed-off-by: Chris Sinjakli <chris@sinjakli.co.uk>