Raise error when grouping key clashes with metric labels in push client
The pushgateway will overwrite metric labels if the same label is used
as part of the grouping key. Other Promethus clients (at least the Go
one) will error if you try to do this, as it can be quite unexpected.
This commit makes us follow suit.

We may find that some users prefer to let those clashes through (i.e.
let the labels from the grouping_key win). If that ends up being the
case, we can introduce a config flag for people to opt into that
behaviour.

Signed-off-by: Chris Sinjakli <chris@sinjakli.co.uk>