Replace Coveralls with SimpleCov
We've been having errors in CI from Coveralls, when trying to upload results.
(As mentioned in [this comment](https://github.com/prometheus/client_ruby/pull/230#issuecomment-864561231))

This error seems to be because our coveralls gem is pretty old and abandoned,
and probably using a TLS version that is no longer supported.

(Reference: https://github.com/lemurheavy/coveralls-ruby/issues/163)

One option recommended in that issue is to switch to a different `coveralls-ruby-reborn` gem.

However, given that we only use `coveralls` to upload results to the cloud,
only so we can have a badge in our README reporting 100%, in the interest of
security, I think i'd rather get rid of `coveralls` altogether, and use
`simplecov` directly instead, which reports the coverage when running the
tests and doesn't upload them anywhere.

Signed-off-by: Daniel Magliola <dmagliola@crystalgears.com>