Handle consecutive path segments containing IDs in collector
Previously, we would fail to strip IDs from paths if they appeared in
consecutive path segments. This is because our regex that looked for IDs
and UUIDs would consume the `/` character that followed them, causing
the next one not to match.

This commit uses a lookahead to match against the `/` without consuming
it.

Signed-off-by: Chris Sinjakli <chris@sinjakli.co.uk>