better to replace the port 5000 with another one since Apple took it for AirPlay
Apple took the port 5000 on the latest MacOS Monterey update, to be used by AirPlay.

Many people are not happy with that. It can be turned on/off easily,
however, a better experience would take that into consideration and use
a different port, considering that MacOS is widely by developers.

Signed-off-by: Hercules Merscher <hlmerscher@gmail.com>