Add `Gauge#set_to_current_time`
I was porting some code from Python to Ruby today, and when I went to
call this method, I realised we don't have it!

My use-case is to track the last time some code successfully ran.

I'm open to discussing the naming, but I've defaulted to the name used
in `client_python` because it seems good to me.

Signed-off-by: Chris Sinjakli <chris@sinjakli.co.uk>