Fix JRuby 9.3 build
The Ruby version targeted by JRuby 9.3 is now too old for Rubygems to
run. Let's pin to the last version that works for now.

When we next have a major version bump for one of the breaking changes
we're planning to make (likely around the store interface), we'll clear
out support for any Ruby versions that have fallen out of security
support.

Signed-off-by: Chris Sinjakli <chris@sinjakli.co.uk>