Add JRuby 9.4 to build matrix
We've had to switch to the main JRuby Docker image as CircleCI have
stopped publishing their own CI-optimised ones with extra tools
installed. It turns out this isn't a problem for us as we're not
dependent on much in the image.

See: https://discuss.circleci.com/t/legacy-convenience-image-deprecation/41034

This commit also switches versions 9.2 and 9.3 to the main JRuby Docker
image, but not 9.1 as that breaks for some inscrutable reason. We'll
drop 9.1 and all other unsupported Rubies in our next major release.

Signed-off-by: Chris Sinjakli <chris@sinjakli.co.uk>