One file per process: move file management from `MetricStore` to `DirectFileStore`
The overall intention here is to have one file per process, instead of one file per process per metric.

We originally went with separate files per metric to make "observe"s faster for multi-threaded processes.
Making the file lock contention be per-metric instead of per-process would theoretically improve performance
by reducing waits on that lock if two threads wanted to observe at the same time.

In practice, however, the result is that we end up with too many files, which causes problems with folks with
large numbers of metrics, and also reading lots of small files instead of a few large ones makes exports slow.

The overall mechanic to achieve this is to add one more special reserved label to all labelsets: ___metric_name
Once the metric name is in the labelset, we can store everything in one file per process.

In the code, this means mostly moving a bunch of the methods dealing with "which files to open" from the per-metric
class (`MetricStore`) to the "global" store class `DirectFileStore`. This allows us to maintain all external interfaces
the same, and tests passing, while "naively" achieving the objective.

In practice, this has a HUGE drawback. When exporting, we're not reading all the (now bigger) files once for every metric and
discarding most of the values we read. I expect this will make exporting outrageously slow.

The answer to this will come on the followup commits, where we'll make the exporter less "elegant" and less "store agnostic",
but hopefully much faster.