Export the DirectFileStore metrics without having to read files multiple times
This is absolutely disgusting, but it's the kind of thing we need to do to get this to work.

There are 2 disgusting parts:
- In the text formatter, we need to read the files directly, and aggregate all the values ourselves
  without going through MetricStore / each metric's individual store. This is pretty horrendous, but
  I think we don't have a way around this.
- Each Metric exposes "values" by reading those values from an internal store. HOWEVER, Summary and Histogram
  do some very clever, very pretty hackery to return a hash instead of a single value, for their buckets, sum, count, etc.

This last thing breaks HORRENDOUSLY in this model, because we need to PASS IN the values we have aggregated, instead of
letting them go down to their internal stores. That's terrible as an interface to expose to our users, which will probably
not be calling this method, but if they do, it's horrible.

We probably need to extract that part for the purposes of formatting, either to the Text formatter, or some intermediary object.
If we do this, either we duplicate the clever code, or users stop having a chance to get those nice values, so we should probably duplicate.

ORRRRR, we completely restructure how everything works. We'd mostly be changing internal interfaces, that most users won't see...
But I don't have a concrete proposal to how to do that.